+++ packages/css-tokenizer/test/test.mjs
@@ -1,16 +1,29 @@
 // Reader
 import './test-reader.mjs';
+
 // Code points
 import './code-points/code-points.mjs';
 import './code-points/ranges.mjs';
+
 // Tokens
 import './token/basic.mjs';
 import './token/comment.mjs';
 import './token/numeric.mjs';
 import './token/url.mjs';
+
 // Complex
 import './complex/at-media-params.mjs';
 import './complex/parse-error.mjs';
+
 // Community
 import './community/bootstrap.mjs';
 import './community/postcss-parser-tests.mjs';
+
+// Stringify
+import './test-stringify.mjs';
+
+// WPT
+import './wpt/cdc-vs-ident-tokens.mjs';
+import './wpt/decimal-points-in-numbers.mjs';
+import './wpt/ident-three-code-points.mjs';
+import './wpt/inclusive-ranges.mjs';