+++ packages/css-parser-algorithms/CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+- Fix: Removes `UnclosedFunctionNode` and `UnclosedSimpleBlockNode`. (breaking)
 - Fix: Do not discard empty items in comma separated lists.
 
 ### 1.0.0 (November 14, 2022)